Dialogs
Novos JDialogFrame foram criados para manter o design do programa

+++ ProjetoFarmacia/src/com/projetofarmacia/dialogs/confirmarPagamentoCartao.java
@@ -10,6 +10,8 @@ import com.projetofarmacia.DAO.VendaDAO;
 import com.projetofarmacia.interfaces.TelaCaixa;
 import com.projetofarmacia.javabeans.Produto;
 import com.projetofarmacia.javabeans.Venda;
+import javax.swing.JTable;
+import javax.swing.table.DefaultTableModel;
 
 /**
  *
@@ -33,14 +35,14 @@ public class confirmarPagamentoCartao extends javax.swing.JDialog {
     private void initComponents() {
 
         btnFinalizar = new javax.swing.JButton();
-        campoTroco = new javax.swing.JTextField();
+        campoNumCartao = new javax.swing.JTextField();
         jLabel1 = new javax.swing.JLabel();
         jLabel2 = new javax.swing.JLabel();
-        campoTroco1 = new javax.swing.JTextField();
+        campoNome = new javax.swing.JTextField();
         jLabel3 = new javax.swing.JLabel();
-        campoTroco2 = new javax.swing.JTextField();
         jLabel4 = new javax.swing.JLabel();
-        campoTroco3 = new javax.swing.JTextField();
+        campoCVV = new javax.swing.JTextField();
+        campoValidade = new javax.swing.JFormattedTextField();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.DISPOSE_ON_CLOSE);
 
@@ -51,9 +53,9 @@ public class confirmarPagamentoCartao extends javax.swing.JDialog {
             }
         });
 
-        campoTroco.addKeyListener(new java.awt.event.KeyAdapter() {
+        campoNumCartao.addKeyListener(new java.awt.event.KeyAdapter() {
             public void keyReleased(java.awt.event.KeyEvent evt) {
-                campoTrocoKeyReleased(evt);
+                campoNumCartaoKeyReleased(evt);
             }
         });
 
@@ -61,27 +63,27 @@ public class confirmarPagamentoCartao extends javax.swing.JDialog {
 
         jLabel2.setText("Insira o nome completo:");
 
-        campoTroco1.addKeyListener(new java.awt.event.KeyAdapter() {
+        campoNome.addKeyListener(new java.awt.event.KeyAdapter() {
             public void keyReleased(java.awt.event.KeyEvent evt) {
-                campoTroco1KeyReleased(evt);
+                campoNomeKeyReleased(evt);
             }
         });
 
         jLabel3.setText("Insira a data de validade:");
 
-        campoTroco2.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                campoTroco2KeyReleased(evt);
-            }
-        });
-
         jLabel4.setText("Insira o CVV:");
 
-        campoTroco3.addKeyListener(new java.awt.event.KeyAdapter() {
+        campoCVV.addKeyListener(new java.awt.event.KeyAdapter() {
             public void keyReleased(java.awt.event.KeyEvent evt) {
-                campoTroco3KeyReleased(evt);
+                campoCVVKeyReleased(evt);
             }
         });
+
+        try {
+            campoValidade.setFormatterFactory(new javax.swing.text.DefaultFormatterFactory(new javax.swing.text.MaskFormatter("##/##")));
+        } catch (java.text.ParseException ex) {
+            ex.printStackTrace();
+        }
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -93,15 +95,15 @@ public class confirmarPagamentoCartao extends javax.swing.JDialog {
                 .addContainerGap())
             .addGroup(layout.createSequentialGroup()
                 .addGap(33, 33, 33)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
                     .addComponent(jLabel4)
-                    .addComponent(campoTroco3, javax.swing.GroupLayout.PREFERRED_SIZE, 187, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(campoTroco, javax.swing.GroupLayout.PREFERRED_SIZE, 187, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(campoCVV, javax.swing.GroupLayout.DEFAULT_SIZE, 187, Short.MAX_VALUE)
+                    .addComponent(campoNumCartao, javax.swing.GroupLayout.DEFAULT_SIZE, 187, Short.MAX_VALUE)
                     .addComponent(jLabel2)
-                    .addComponent(campoTroco1, javax.swing.GroupLayout.PREFERRED_SIZE, 187, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(campoNome, javax.swing.GroupLayout.DEFAULT_SIZE, 187, Short.MAX_VALUE)
                     .addComponent(jLabel1)
                     .addComponent(jLabel3)
-                    .addComponent(campoTroco2, javax.swing.GroupLayout.PREFERRED_SIZE, 187, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(campoValidade))
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
@@ -110,19 +112,19 @@ public class confirmarPagamentoCartao extends javax.swing.JDialog {
                 .addGap(29, 29, 29)
                 .addComponent(jLabel1)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(campoTroco, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(campoNumCartao, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jLabel2)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(campoTroco1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(campoNome, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(18, 18, 18)
                 .addComponent(jLabel3)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(campoTroco2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(campoValidade, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jLabel4)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(campoTroco3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(campoCVV, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 52, Short.MAX_VALUE)
                 .addComponent(btnFinalizar)
                 .addContainerGap())
@@ -131,36 +133,43 @@ public class confirmarPagamentoCartao extends javax.swing.JDialog {
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
-    private void campoTrocoKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_campoTrocoKeyReleased
+    private void campoNumCartaoKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_campoNumCartaoKeyReleased
        
-    }//GEN-LAST:event_campoTrocoKeyReleased
+    }//GEN-LAST:event_campoNumCartaoKeyReleased
 
     private void btnFinalizarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnFinalizarActionPerformed
-        Venda obj = new Venda();
-        VendaDAO dao = new VendaDAO();
-        Produto p = new Produto();
-        ProdutoDAO pdao = new ProdutoDAO();
-        for (int i = 0; i < TelaCaixa.tabelaProduto.getRowCount(); i++) {
-            obj.setStatus("FINALIZADO");
-            pdao.removeQuantidade(p);
-            dao.finalizarVenda(obj);
-            
+        if (campoNome.getText().equals("") || campoNumCartao.getText().equals("")
+            || campoValidade.getText().equals("") || campoCVV.getText().equals("")) {
+            new camposVazios(null, true).setVisible(true);
+        } else {
+            Venda obj = new Venda();
+            VendaDAO dao = new VendaDAO();
+            Produto p = new Produto();
+            ProdutoDAO pdao = new ProdutoDAO();
+            for (int i = 0; i < TelaCaixa.tabelaProduto.getRowCount(); i++) {
+                obj.setStatus("FINALIZADO");
+                pdao.removeQuantidade(p);
+                dao.finalizarVenda(obj);
+            }
+            limparTabela(TelaCaixa.tabelaProduto);
         }
         
         
     }//GEN-LAST:event_btnFinalizarActionPerformed
 
-    private void campoTroco1KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_campoTroco1KeyReleased
+    private void limparTabela(JTable table) {
+        DefaultTableModel tabela = (DefaultTableModel) table.getModel();
+        for (int i = tabela.getRowCount() -1; i >= 0 ; i--) {
+            tabela.removeRow(i);
+        }
+    }
+    private void campoNomeKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_campoNomeKeyReleased
         // TODO add your handling code here:
-    }//GEN-LAST:event_campoTroco1KeyReleased
+    }//GEN-LAST:event_campoNomeKeyReleased
 
-    private void campoTroco2KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_campoTroco2KeyReleased
+    private void campoCVVKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_campoCVVKeyReleased
         // TODO add your handling code here:
-    }//GEN-LAST:event_campoTroco2KeyReleased
-
-    private void campoTroco3KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_campoTroco3KeyReleased
-        // TODO add your handling code here:
-    }//GEN-LAST:event_campoTroco3KeyReleased
+    }//GEN-LAST:event_campoCVVKeyReleased
 
     /**
      * @param args the command line arguments
@@ -207,10 +216,10 @@ public class confirmarPagamentoCartao extends javax.swing.JDialog {
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     public static javax.swing.JButton btnFinalizar;
-    private javax.swing.JTextField campoTroco;
-    private javax.swing.JTextField campoTroco1;
-    private javax.swing.JTextField campoTroco2;
-    private javax.swing.JTextField campoTroco3;
+    private javax.swing.JTextField campoCVV;
+    private javax.swing.JTextField campoNome;
+    private javax.swing.JTextField campoNumCartao;
+    private javax.swing.JFormattedTextField campoValidade;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
